requirements.txt
django==1.11.*
djangorestframework==3.7.*
celery==4.*
boto3==1.5.*
structlog==18.*
requests==2.18.*
